ja3s-string: move unittests to tests/
